Refactor: Move tokenization and model setup utilities to separate modules
- Extracted tokenization logic (setup_tokenizer, tokenize_datasets) into tokenization_utils.py
- Extracted model configuration logic (setup_model_and_config) into model_utils.py
- Updated train_tiny_stories.py to use new utility modules
- Improved code organization and modularity for easier maintenance and extension
