+++ train_tiny_stories.py
@@ -80,7 +80,10 @@ import sys
 import time
 
 # Import the new CacheManager
-from utils import CacheManager
+from cache_utils import CacheManager
+from dataset_utils import load_tinystories_dataset, create_dataset_subsets
+from tokenization_utils import setup_tokenizer, tokenize_datasets
+from model_utils import setup_model_and_config
 
 # =====================
 # Prompt Selection Logic
@@ -117,229 +120,8 @@ def determine_generation_prompt(args: argparse.Namespace) -> str:
         return "Once upon a time, there was a little"
 
 # =====================
-# Dataset utilities
-# =====================
-
-
-def load_tinystories_dataset() -> Any:
-    """
-    Load the TinyStories dataset from HuggingFace Datasets.
-
-    TinyStories is a dataset of short stories (~3-5 sentences) generated by GPT-3.5/4,
-    designed to be simple enough for small models to learn basic language structure.
-
-    Returns:
-        The loaded dataset object.
-    """
-    print("Loading TinyStories dataset...")
-    dataset = load_dataset("roneneldan/TinyStories")
-    print(f"Full TinyStories dataset size: {len(dataset['train'])} training examples")
-    return dataset
-
-
-def create_dataset_subsets(
-    dataset: Any,
-    num_examples: int,
-    filter_word: Optional[str],
-    no_cache: bool,
-    cache_manager: CacheManager,
-) -> Tuple[Any, Any, str]:
-    """
-    Create (and cache) training and validation subsets from the TinyStories dataset.
-    Args:
-        dataset: The loaded TinyStories dataset.
-        num_examples: Number of training examples to use.
-        filter_word: Optional word to filter examples.
-        no_cache: If True, disables caching.
-        cache_manager: Instance of CacheManager for handling caching.
-    Returns:
-        (train_subset, val_subset, dataset_cache_key)
-    """
-    dataset_params = {"num_examples": num_examples, "filter_word": filter_word}
-    
-    def compute_subsets():
-        print(f"Creating dataset subsets with {num_examples} examples...")
-        subset_size = num_examples
-
-        if filter_word:
-            print(f"Filtering dataset to include only examples with the word: '{filter_word}'")
-
-            def filter_function(example):
-                return filter_word.lower() in example["text"].lower()
-
-            filtered_dataset = dataset["train"].filter(
-                filter_function, desc=f"Filtering for '{filter_word}'"
-            )
-            print(f"Found {len(filtered_dataset)} examples containing '{filter_word}'")
-            if len(filtered_dataset) > subset_size:
-                train_subset = filtered_dataset.select(range(subset_size))
-            else:
-                train_subset = filtered_dataset
-                print(
-                    f"Warning: Only {len(train_subset)} examples contain '{filter_word}', less than requested {subset_size}"
-                )
-        else:
-            train_subset = dataset["train"].select(range(subset_size))
-            
-        print(f"Using subset size: {len(train_subset)} training examples")
-        val_size = int(len(train_subset) * 0.1)
-        val_start_idx = subset_size if not filter_word else 0
-        val_subset = dataset["train"].select(range(val_start_idx, val_start_idx + val_size))
-        print(f"Created validation set with {len(val_subset)} examples")
-        
-        return {"train_subset": train_subset, "val_subset": val_subset}
-    
-    # Use the new helper method
-    cached_data, dataset_cache_key = cache_manager.get_or_compute(
-        dataset_params, "dataset_subsets", compute_subsets, no_cache
-    )
-    
-    return cached_data["train_subset"], cached_data["val_subset"], dataset_cache_key
-
-
-# =====================
-# Tokenization utilities
-# =====================
-
-
-def setup_tokenizer() -> GPT2Tokenizer:
-    """
-    Load and configure the GPT2 tokenizer, adding a pad token if needed.
-    Returns:
-        The configured tokenizer.
-    """
-    print("Setting up the tokenizer...")
-    tokenizer = GPT2Tokenizer.from_pretrained("gpt2")
-    if tokenizer.pad_token is None:
-        tokenizer.pad_token = tokenizer.eos_token
-        print("Added padding token (same as EOS token)")
-    return tokenizer
-
-
-def tokenize_datasets(
-    train_subset: Any,
-    val_subset: Any,
-    tokenizer: GPT2Tokenizer,
-    dataset_cache_key: str,
-    no_cache: bool,
-    max_length: int,
-    cache_manager: CacheManager,
-) -> Tuple[Any, Any, str]:
-    """
-    Tokenize the training and validation datasets, with caching.
-
-    Why Tokenize?
-    -------------
-    LLMs (Large Language Models) don't process raw text directly.
-    They require numerical input. Tokenization is the process of converting
-    text into a sequence of numerical IDs (tokens).
-    - Vocabulary: Each unique ID corresponds to a word or sub-word piece
-      defined in the tokenizer's vocabulary.
-    - Padding: Ensures all sequences in a batch have the same length, which is
-      often required for efficient processing by the model. Shorter sequences
-      are padded with a special token.
-    - Truncation: Cuts off sequences that are longer than the maximum length
-      the model can handle.
-
-    Args:
-        train_subset: Training dataset subset.
-        val_subset: Validation dataset subset.
-        tokenizer: The tokenizer to use.
-        dataset_cache_key: Cache key for the dataset subset.
-        no_cache: If True, disables caching.
-        max_length: Maximum sequence length for tokenization.
-        cache_manager: Instance of CacheManager for handling caching.
-    Returns:
-        (tokenized_train, tokenized_val, tokenization_cache_key)
-    """
-    def tokenize_function(examples: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Tokenize the text examples with padding and truncation to a fixed length.
-        Args:
-            examples: Dictionary of text examples
-        Returns:
-            Dictionary with tokenized inputs (input_ids, attention_mask)
-        """
-        return tokenizer(
-            examples["text"],
-            padding="max_length",
-            truncation=True,
-            max_length=max_length,
-        )
-
-    tokenization_params = {
-        "dataset_cache_key": dataset_cache_key,
-        "max_length": max_length,
-        "tokenizer_name": "gpt2",
-    }
-    
-    def compute_tokenization():
-        print("Tokenizing the dataset (this might take a while)...")
-        tokenized_train = train_subset.map(
-            tokenize_function, batched=True, desc="Tokenizing training data"
-        )
-        tokenized_val = val_subset.map(
-            tokenize_function, batched=True, desc="Tokenizing validation data"
-        )
-        return {"tokenized_train": tokenized_train, "tokenized_val": tokenized_val}
-    
-    # Use the new helper method
-    cached_data, tokenization_cache_key = cache_manager.get_or_compute(
-        tokenization_params, "tokenized_datasets", compute_tokenization, no_cache
-    )
-    
-    return cached_data["tokenized_train"], cached_data["tokenized_val"], tokenization_cache_key
-
-
-# =====================
 # Model utilities
 # =====================
-
-
-def setup_model_and_config(
-    tokenizer: GPT2Tokenizer, dimensions: int, layers: int, heads: int, max_length: int
-) -> Tuple[GPT2Config, Dict[str, Any]]:
-    """
-    Create a GPT2Config and model parameter dictionary for a small LLM.
-
-    This defines the architecture of our transformer model.
-    - n_embd (dimensions): The size of the vector representing each token (embedding dimension).
-    - n_layer (layers): The number of transformer blocks stacked on top of each other. More layers = deeper model.
-    - n_head (heads): The number of attention heads in the multi-head attention mechanism. Allows the model to focus on different parts of the input simultaneously. Must divide n_embd evenly.
-    - n_positions (max_length): The maximum sequence length the model can process.
-
-    Args:
-        tokenizer: The tokenizer to use.
-        dimensions: Embedding dimension size.
-        layers: Number of transformer layers.
-        heads: Number of attention heads.
-        max_length: Maximum sequence length for the model.
-    Returns:
-        (config, model_params)
-    """
-    model_params = {
-        "tokenizer_vocab_size": tokenizer.vocab_size,
-        "n_positions": max_length,
-        "n_embd": dimensions,
-        "n_layer": layers,
-        "n_head": heads, # Use the heads argument
-        "pad_token_id": tokenizer.pad_token_id,
-        "eos_token_id": tokenizer.eos_token_id,
-    }
-    print("Creating a small GPT-2 model configuration...")
-    config = GPT2Config(
-        vocab_size=model_params["tokenizer_vocab_size"],
-        n_positions=model_params["n_positions"],
-        n_embd=model_params["n_embd"],
-        n_layer=model_params["n_layer"],
-        n_head=model_params["n_head"], # Use heads here
-        pad_token_id=model_params["pad_token_id"],
-        eos_token_id=model_params["eos_token_id"],
-    )
-    print(
-        f"Model parameters: layers={config.n_layer}, heads={config.n_head}, embd_dim={config.n_embd}"
-    )
-    return config, model_params
 
 
 # =====================
